Move file functions out to utility file
